DHFPROD-4951: UI is now using new step and flow endpoints
This intentionally is not trying to make the uploading of files work for ingestion steps, nor are they runnable, as there are separate stories for changing how those features work. 

The tests are failing in Bench.test.tsx, can't figure out why - manual testing of the Bench seems to confirm that everything is working properly. 

Otherwise, the UI changes were fairly simple to make - the only interesting thing was modifying flows.tsx to take advantage of getting all the necessary data from /api/flows instead of having to get all the ingestion and mapping steps too. 

Removed the functions in artifacts/core for settings, along with related tests, as we no longer need a concept of a separate settings document. Will remove the rest later, once we know everything here is working.